Add useNote and useNoteReplies
